@@ -13,9 +13,7 @@ class PostFeedController extends Controller
      */
     public function index(): Response
     {
-        $posts = Cache::remember('feed-posts', now()->addHour(), function () {
-            return Post::latest()->limit(20)->get();
-        });
+        $posts = Cache::remember('feed-posts', now()->addHour(), fn () => Post::latest()->limit(20)->get());
 
         return response()->view('posts_feed.index', [
             'posts' => $posts
