@@ -39,8 +39,6 @@ class TelescopeServiceProvider extends TelescopeApplicationServiceProvider
      */
     protected function gate()
     {
-        Gate::define('viewTelescope', function ($user) {
-            return $user->isAdmin();
-        });
+        Gate::define('viewTelescope', fn ($user) => $user->isAdmin());
     }
 }
