@@ -8,11 +8,7 @@ use Faker\Generator;
 $factory->define(Comment::class, function (Generator $faker) {
     return [
         'content' => $faker->paragraph,
-        'author_id' => function () {
-            return factory(User::class)->create()->id;
-        },
-        'post_id' => function () {
-            return factory(Post::class)->create()->id;
-        }
+        'author_id' => fn () => factory(User::class)->create()->id,
+        'post_id' => fn () => factory(Post::class)->create()->id
     ];
 });
