@@ -65,7 +65,7 @@ return [
             'servers' => [
                 [
                     'host' => env('MEMCACHED_HOST', '127.0.0.1'),
-                    'port' => env('MEMCACHED_PORT', 11211),
+                    'port' => env('MEMCACHED_PORT', 11_211),
                     'weight' => 100,
                 ],
             ],
